feat: handle notification for like follow and comment

+++ src/app/api/v1/posts/[postId]/reactions/route.ts
@@ -47,21 +47,46 @@ export const POST = async (req: NextRequest, { params: { postId } }: { params: {
   try {
     const { user } = await validateRequest()
     if (!user) return NextResponse.json({ error: "unauthorized" }, { status: 401 })
-    const post = await prisma.like.upsert({
+    const post = await prisma.post.findUnique({
       where: {
-        userId_postId: {
+        id: postId,
+      },
+      select: {
+        userId: true,
+      },
+    })
+    if (!post) return NextResponse.json({ error: "Post not found" }, { status: 404 })
+
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: user.id,
+            postId,
+          },
+        },
+        create: {
           userId: user.id,
           postId,
         },
-      },
-      create: {
-        userId: user.id,
-        postId,
-      },
-      update: {},
-    })
-    console.log("🚀 ~ POST ~ post:", post)
-    return new Response()
+        update: {},
+      }),
+
+      ...(user.id === postId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: user.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ])
+
+    return NextResponse.json({ success: true })
   } catch (err) {
     console.log("🚀 ~ POST ~ err:", err)
     return NextResponse.json({ error: "Internal server error" }, { status: 500 })
@@ -71,14 +96,35 @@ export const DELETE = async (req: NextRequest, { params: { postId } }: { params:
   try {
     const { user } = await validateRequest()
     if (!user) return NextResponse.json({ error: "unauthorized" }, { status: 401 })
-    const post = await prisma.like.deleteMany({
+    const post = await prisma.post.findUnique({
       where: {
-        userId: user.id,
-        postId,
+        id: postId,
+      },
+      select: {
+        userId: true,
       },
     })
-    console.log("🚀 ~ DELETE ~ post:", post)
-    return new Response()
+    if (!post) return NextResponse.json({ error: "Post not found" }, { status: 404 })
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: user.id,
+          postId,
+        },
+      }),
+
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: user.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ])
+
+    return NextResponse.json({ success: true })
   } catch (err) {
     console.log("🚀 ~ POST ~ err:", err)
     return NextResponse.json({ error: "Internal server error" }, { status: 500 })
